Numerous fixes to support CI across platforms.
The main discovery trying to get this running on macOS was that the
PythonBuildStandalone Python distributtions can not be counted on for
building platform specific distributions*. This means macOS aarch64
Pants installs tend to fail trying build distributions Pants depends on
like psutil and setproctitle. To avoid this issue, there is a set of
changes to use the binaries.pantsbuild.org pre-built wheels via a
--find-links repo for all installs.

In order to support a fully statically linked `scie-pants` on Linux, the
package crate builds steps are broken up and re-worked as subcommands.
This arrangement allows for building the tools.pex on Ubuntu (glibc),
then the scie-pants scie in an alpine image where it gets fully
statically linked, and then running the integration tests back in
Ubuntu.

* See: https://github.com/indygreg/python-build-standalone/blob/main/docs/quirks.rst#references-to-build-time-paths