Achieve parity with one_step_setup and then some. (#27)
Pants will now configure `pants_version` to be latest stable if
`pants.toml` is present but `pants_version` is not. Further, it will
establish a `pants.toml` if none is present. If the CWD is nested in
a `.git` dir, the `pants.toml` is created there instead.

To support configuration in these cases, break out a `configure`
binding to handle prompting and version selection separate from the
`install` binding.