Improve Pants help output / CLI fidelity.
Previously scie-pants over-rode Pants help defaulting to `./pants`
in help text by passing `--pants-bin-name` as the absolute path of
scie-pants. Now scie-pants passes the exact argv0 string it was
invoked with and it does so using the `PANTS_BIN_NAME` env var instead
of `--pants-bin-name` to support users over-riding this.

Closes #111