Update scie-pants to scie-jump 0.11.1. (#199)
This pulls in a fix for handling of regex env var removal when the
ambient env contains non-utf8 entries. Out of the 4 ways Pants can be
run, threeare still broken, but these breaks are inside Pants itself.
Two breaks occur when using pantsd. There is one in the native client
and one in the `PyNailgunClient`. Another occurs using `--no-pants`
with the native client. For more about those issues, see:
https://github.com/pantsbuild/pants/issues/19199

Fixes #198