Grab latest version from GitHub Release tag (#225)
As discussed in the latest Pants meetup and [this
proposal](https://github.com/pantsbuild/pants/discussions/19443), we're
going to move off of PyPI for distribution, and instead leverage GitHub
Releases.

Some reasons given for this pivot:
- People ought not to `pip install pantsbuild.pants`
- We ought not to continue eating up PyPI space, which we're already
hitting the ceiling for
- We plan on having `scie-pants` use a PEX in the future, and PyPI
wouldn't support that workflow
- We want to abstract away, as much is reasonable, that Pants is
implemented in Python. And treat it like a black box, where possible.

Currently the latest version is expected to be set by hand. Although
that's unlikely to be wrong, we're also changing to have the "latest"
marker be [automated](https://github.com/pantsbuild/pants/pull/19444).
Therefore we should be able to rely on it with confidence.

Tested and both before/after yielded the same string (with the GitHub
URL performing slightly faster)