Propagate scie-pants' version into the pants invocation (#246)
This sets the `SCIE_PANTS_VERSION` env var when invoking pants, so that
pants can read it and encourage users to upgrade if they need to.

For example, 2.17 will be the last release that supports the PyPI-based
distribution model and Pants 2.18 will require scie-pants 0.9 or
greater. So, in https://github.com/pantsbuild/pants/issues/19600, we're
planning to have pants 2.17 encourage users to upgrade `scie-pants` if
it's too old.

This PR preps for the release too, by bumping the version and adding a
changelog entry. As part of this, I noticed the 0.9.0 release was
accidentally configured as 0.9.1 in Cargo.toml, so updated the changelog
to mention this.