Pass --python-repos-find-links always, and deprecate PANTS_SHA (#250)
Currently, the `scie-pants` launcher sometimes needs to pass an
additional argument to the underlying pants invocation, to be able to
support `PANTS_SHA` and its custom wheel location. Effectively turning
an invocation like `pants goal --arg` into `pants
--pants-repos-find-links=+["..."] goal --arg` when appropriate:
https://github.com/pantsbuild/scie-pants/blob/0a204c5ac816ad24d74b55ecf7424397a07b3c2a/package/scie-pants.toml#L57-L60

However, when not using `PANTS_SHA`, the pants command becomes `pants ""
goal --arg`, and the spurious empty string results in some commands
failing, e.g. `pants "" tailor --check` thinks `--check` is a global
argument, and this is ambiguous.

To fix this issue, this PR does two things:

1. avoid passing an empty string, via a no-op workaround: always set the
`PANTS_SHA_FIND_LINKS` variable (to a no-op like
`--python-repos-find-links=-[]`), so that scie-pants never passes an
empty variable to the pants command, since that makes commands like
`pants tailor --check ::` explode:
2. deprecates `PANTS_SHA`, so that the hack above doesn't feel so bad:
we're on the way to removing the need for `PANTS_SHA_FIND_LINKS`
entirely, thus sidestepping the limitation and the need for the
workaround (2)

This could also be fixed via https://github.com/a-scie/jump/issues/130.

Fixes #249