Upgrade to Pants 2.16 for Python code (#259)
Pants 2.16 has been around for a while, let's have scie-pants use it.

This requires adjusting the tools due to the changes to tool lockfiles
in 2.16. I took the approach of removing those custom lockfiles: this is
a Pantsbuild repo, so hopefully the default ones are appropriate (if
not, why are we offering them to others 😅 )!