Pin example-django to a specific commit in tests (#292)
This pins tests to run against a specific commit of `example-django`, a
commit that has the versions expected. In particular, `example-django`
recently been upgraded to 2.17.0
https://github.com/pantsbuild/example-django/commit/b94e0b519408205961b28077f60cbc26776622a6
which has some breaking changes, that stop 2.12.1 from running on
2.17.0, and thus the test that attempts to run with `PANTS_VERSION` to
downgrade was failing.

This pins to most recent working version: the parent of the breaking
commit.