Fix PBS linking against libcrypt.so.1 for real. (#344)
The prior PBS upgrade was to a version that claimed to have this fixed,
but it was only fixed for 3.11 and not actually fixed for `3.{8,9,10}`.
This time we upgrade again to

https://github.com/indygreg/python-build-standalone/releases/tag/20240107
which notes this is actually now fixed for those Pythons. A manual test
of scie-pants 0.10.5 inside a fedora:37 container confirms the current
broken status and mounting in a freshly build scie-pants using this
change confirms a fix.

Fixes #52