Fix CI: Use a tmpdir SCIE_BASE to force bootstrapping, incorporate Python path into cache key (#385)
There were two issues with CI:

- It seems like the test that verifies silence on stdout wasn't always
doing the required bootstrap. This makes the bootstrap always happen by
forcing it to be starting with an empty cache (`SCIE_BASE`) by setting
that to a fresh temp dir.
- The ubuntu runner was failing to run the `test_pants_source_mode`
test, because the venv from the restored cache had an symlink to
`/opt/hostedtoolcache/Python/3.9.18/x64/bin/python` (NB version 3.9.18),
but the runners have updated to 3.9.19. Fixes #270.