Use Pants'-managed runs-on runners for Linux aarch64 CI (#427)
This replaces our use of Circle CI's arm64 Linux runners for both
testing and releasing with the runs-on runners that Pantsbuild
configures (and pays for) in AWS, via https://runs-on.com

This has a few benefits:

- single tool to learn for configuring CI
- single source of credentials (currently the release build is failing
due to a token seemingly being invalid, but it's not obvious to me how
to fix this)
- more atomicity, e.g. instead of the release needing to trigger a
separate Circle CI build that runs in the background, we can just use
the normal GitHub "wait for another job" rules

It also has downsides: we're going from free resources to paid. However,
the incremental cost for the scie-pants repo should be quite small,
compared to the main pants repo:

- this repo is low volume (over the last 12 months we've done about 220
CI runs total)
- things run quickly (in this PR, ~7 minutes without a cache, ~4 minutes
with)

I estimated cost based on US$0.1344/hour for a on-demand t4g.xlarge 4
core instance in us-east-1 (I'm not sure this is our exact
configuration). Prices in practice may be cheaper, because Runs-On
attempts to use [spot
instances](https://runs-on.com/configuration/spot-instances/)).

- per build, we're looking at US$0.0090 (with cache) to US$0.016
(without cache)
- per month, we're looking at ~20 builds, so US$0.18 to US$0.32 (but
likely cheaper given this is on-demand costs)

If this starts adding up a lot, we can likely optimise. For instance:
optimise spot-instance selection for lowest cost not "PCO", and use more
caching (whether that's coarse-grained `actions/cache` or fine-grained
`sccache`).

Fixes #422, fixes #430