Build a draft release and then publish once complete (#432)
Instead of publishing partial releases (even for a few minutes) as the
matrix of builds adds their artefacts, create a draft release that we
then publish in once the build has finished.

Background: the build process involves some set-up, then a matrix of
building for different platforms, then some finalisation. This change is
moving the "publish the release" from the matrix jobs to the
finalisation job.


![image](https://github.com/user-attachments/assets/ec8d2161-6fe3-471c-bdc7-aa1f887f1cf4)


This has a few benefits:

- Atomic publishing: a release is only published once it has artifacts
for all platforms. Before this change, if a builder fails for one
platform, we'll have published a "latest" release visible to
`get-pants.sh`, and so running that on the broken platform will fail.
(Example: #430.)
- Reduced duplicate work: rather than have each builder in the matrix
compute changelogs etc., just do it once as part of the publishing. This
is more conceptually helpful, than an optimisation, since it only takes
a few seconds to prep the changelog.